feat: enhance error handling and type definitions across various components

+++ backend/financeapp-core/src/main/java/com/financeapp/core/service/DashboardService.java
@@ -68,7 +68,7 @@ public class DashboardService {
                 .totalExpenses(totalExpenses)
                 .savingsRate(savingsRate)
                 .transactionCount(transactions.size())
-                .categoryCount(categoryRepository.countByUserId(userId))
+                .categoryCount((int) categoryRepository.countByUserId(userId))
                 .build();
     }
